"add registry interface"
